refactor(api): Rename file directory and update resource paths in console API
