+++ api/controllers/console/__init__.py
@@ -2,16 +2,16 @@ from flask import Blueprint
 
 from libs.external_api import ExternalApi
 
-from .file import FileApi, FilePreviewApi, FileSupportTypeApi
+from .files import FileApi, FilePreviewApi, FileSupportTypeApi
 from .remote_files import RemoteFileInfoApi, RemoteFileUploadApi
 
 bp = Blueprint("console", __name__, url_prefix="/console/api")
 api = ExternalApi(bp)
 
 # File
-api.add_resource(FileApi, "/file/upload")
-api.add_resource(FilePreviewApi, "/file/<uuid:file_id>/preview")
-api.add_resource(FileSupportTypeApi, "/file/support-type")
+api.add_resource(FileApi, "/files/upload")
+api.add_resource(FilePreviewApi, "/files/<uuid:file_id>/preview")
+api.add_resource(FileSupportTypeApi, "/files/support-type")
 
 # Remote files
 api.add_resource(RemoteFileInfoApi, "/remote-files/<path:url>")
